+++ scripts/BuildLinuxArm64.ps1
@@ -1,0 +1,8 @@
+param()
+
+[string] $Product = "Mzinga"
+[string] $Target = "LinuxArm64"
+
+& "$PSScriptRoot\Build.ps1" -Product $Product -Target $Target -BuildArgs "-target:Publish -p:RuntimeIdentifier=linux-arm64 -p:PublishSingleFile=true -p:PublishTrimmed=true -p:IncludeAllContentForSelfExtract=true -p:EnableCompressionInSingleFile=true"
+
+& "$PSScriptRoot\TarRelease.ps1" -Product $Product -Target $Target